FWF-3460:[Feature]- Edit Page New Design & Universal scrolling
